refactor(frontend): split SQL console support
